added delete button

+++ api/routes/deleteTrack.js
@@ -3,7 +3,7 @@
 const sendStatus = require("../helpers.js").sendStatus;
 const dbQuery = require("../db.js").dbQuery;
 const upload = require("../forms.js").upload;
-const IMAGE_PATH = require("../globals.js").IMAGE_PATH;
+const ALBUM_PATH = require("../globals.js").ALBUM_PATH;
 
 const express = require("express");
 const app = express();
@@ -11,13 +11,17 @@ const path = require("path");
 const fs = require("fs");
 
 module.exports = app.post('/', upload.none(), async (req, res) => {
+	if (!req.authed) {
+		sendStatus(req, res, 401, "error.forbidden");
+		return;
+	}
+
 	let userTrackToDelete = await dbQuery("SELECT id, track_id FROM user_tracks JOIN users ON user_tracks.uid = users.uid WHERE users.uid = ? AND id = ?", [req.profile.uid, req.body.id]);
 	if (!userTrackToDelete.length) {
 		sendStatus(req, res, 404, "error.no_track_ownership");
 		return;
 	}
-	let deletion = await dbQuery("DELETE FROM user_tracks WHERE id = ?", [userTrackToDelete[0].id]);
-
+	await dbQuery("DELETE FROM user_tracks WHERE id = ?", [userTrackToDelete[0].id]);
 	sendStatus(req, res, 200, "success.deletion_ok");
 
 	let isTrackPopulated = (await dbQuery("SELECT 1 FROM user_tracks WHERE track_id = ?", [userTrackToDelete[0].track_id])).length;
@@ -25,7 +29,7 @@ module.exports = app.post('/', upload.none(), async (req, res) => {
 	if (!isTrackPopulated) {
 		let track = await dbQuery("SELECT id, image FROM tracks WHERE id = ?", [userTrackToDelete[0].track_id]);
 		let trackToDelete = await dbQuery("DELETE FROM tracks WHERE id = ?", [track[0].id]);
-		fs.rm(path.join(__dirname, IMAGE_PATH + "album_covers/" + track[0].image), e => { if (e) console.log(e)});
+		fs.rm(path.join(__dirname, `${ALBUM_PATH}/${track[0].image}`), e => { if (e) console.log(e)});
 	}
 });
